+++ back-end/src/controllers/v1/locationController.ts
@@ -81,9 +81,15 @@ export const deleteLocation = async (
 		const { userId, locationId } = getRequestVariables(req, true);
 		if (!(await hasLocationPermission(userId, locationId, Location.EDIT_ALL)))
 			return res.status(403).json({ message: "Forbidden" });
+
 		const sql = `DELETE FROM locations WHERE id = ?`;
-		await database.query(sql, [locationId], userId);
 		return res.status(200).json({ message: "Location deleted successfully" });
+		const result = await database.query(sql, [locationId], userId);
+
+		if (!result) {
+			return res.status(500).json({ message: "Internal server error" });
+		}
+
 		await database.query(
 			`DELETE FROM location_response WHERE location_id IN (SELECT id FROM event_locations WHERE location_id = ?)`,
 			[locationId],
@@ -112,8 +118,12 @@ export const createLocation = async (
 		if (!locationName)
 			return res.status(400).json({ message: "Missing location name" });
 		const sql = `INSERT INTO locations (creator_user, name) VALUES (?, ?)`;
-		await database.query(sql, [userId, locationName], userId);
 		return res.status(201).json({ message: "Location created successfully" });
+		const result = await database.query(sql, [userId, locationName], userId);
+
+		if (!result || result.rows.length === 0) {
+			return res.status(500).json({ message: "Internal server error" });
+		}
 	} catch (err) {
 		next(err);
 	}
@@ -134,8 +144,15 @@ export const updateLocation = async (
 			: null;
 		if (!locationName)
 			return res.status(400).json({ message: "Missing location name" });
-		await database.query(sql, [locationName, locationId], userId);
-		return res.status(200).json({ message: "Location updated successfully" });
+		const result = await database.query(
+			sql,
+			[locationName, locationId],
+			userId,
+		);
+		if (!result) {
+			return res.status(500).json({ message: "Internal server error" });
+		}
+		return res.status(204).json();
 	} catch (err) {
 		next(err);
 	}
@@ -160,8 +177,15 @@ export const updateFullLocation = async (
 		await ifMatchValidator(req, `SELECT * FROM locations WHERE id = ?`, [
 			locationId,
 		]);
-		await database.query(sql, [locationName, locationId], userId);
 		return res.status(200).json({ message: "Location updated successfully" });
+		const result = await database.query(
+			sql,
+			[locationName, locationId],
+			userId,
+		);
+		if (!result) {
+			return res.status(500).json({ message: "Internal server error" });
+		}
 	} catch (err) {
 		next(err);
 	}
@@ -255,10 +279,13 @@ export const createEventLocation = async (
 			return res.status(403).json({ message: "Forbidden" });
 		}
 		const sql = `INSERT INTO event_locations (event_id, location_id) VALUES (?, ?)`;
-		await database.query(sql, [eventId, locationId], userId);
 		return res
 			.status(201)
 			.json({ message: "Event location created successfully" });
+		const result = await database.query(sql, [eventId, locationId], userId);
+		if (!result || result.rows.length === 0) {
+			return res.status(500).json({ message: "Internal server error" });
+		}
 	} catch (err) {
 		next(err);
 	}
@@ -282,10 +309,14 @@ export const deleteEventLocation = async (
 				return res.status(403).json({ message: "Forbidden" });
 			}
 			const sql = `DELETE FROM event_locations WHERE event_id = ? AND location_id = ?`;
-			await database.query(sql, [eventId, locationId], userId);
 			return res
 				.status(201)
 				.json({ message: "Event location deleted successfully" });
+			const result = await database.query(sql, [eventId, locationId], userId);
+
+			if (!result) {
+				return res.status(500).json({ message: "Internal server error" });
+			}
 
 			await database.query(
 				`DELETE FROM location_response WHERE location_id = ?`,
